extract invalidateAPIKey fns into shared package
