+++ packages/shared/src/server/index.ts
@@ -11,6 +11,7 @@ export * from "./services/traces-ui-table-service";
 export * from "./services/InMemoryFilterService";
 export * from "./evalJobConfigCache";
 export * from "./auth/apiKeys";
+export * from "./auth/invalidateApiKeys";
 export * from "./auth/customSsoProvider";
 export * from "./auth/gitHubEnterpriseProvider";
 export * from "./llm/fetchLLMCompletion";
